fix: add missing logic to prevent user from donation when the receive amount cannot be determined
