+++ dws/src/components/metamask/donatev2/ConnectBtn.tsx
@@ -3,12 +3,14 @@ import { MouseEvent, ReactElement } from "react";
 
 interface ConnectButtonProps {
   account: Nullable<string>;
+  disabled: boolean;
   handleConnect: (event: MouseEvent<HTMLButtonElement>) => void;
   handleDonate: (event: MouseEvent<HTMLButtonElement>) => void;
 }
 
 const ConnectButton = ({
   account,
+  disabled,
   handleConnect,
   handleDonate,
 }: ConnectButtonProps): ReactElement<ConnectButtonProps> => {
@@ -19,8 +21,9 @@ const ConnectButton = ({
           Connected to {account}
         </div>
         <button
-          className="w-full bg-cred border-2 border-black rounded-2xl py-2"
+          className="w-full bg-cred border-2 border-black rounded-2xl py-2 disabled:opacity-60 disabled:cursor-not-allowed"
           onClick={handleDonate}
+          disabled={disabled}
         >
           <div className="text-xl leading-loose tracking-wider text-gray-100">
             Donate
